requirements.txt
# App
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.23

# Testing
pytest==8.3.5
pytest-cov==5.0.0
factory-boy==3.3.1
Faker==25.8.0

# Utilities
python-dotenv==1.0.0